allow to specify unfolding to full height via special value
JSONlim::Height in the API, 'height' or '∞' in the command line.
So far the API used nil as such a special value; now stop supporting
that due the ambiguity of a nil valued parameter (as nil value can
be expected to mean "I want the default value", but also "I want
to have no limit").